Create web3token.sol
The code defines a smart contract named "W3BVIII" written in Solidity programming language. The contract implements a token contract with total supply, decimal, name and symbol. The contract uses state variables to store the token details like owner, name, symbol, decimal and totalSupply. The contract uses mappings to store the balance of each address and the amount that a user is allowed to spend by another user. The contract implements various functions like name_(), symbol_(), _decimal(), _totalSupply(), _balanceOf() to get the respective details of the token. The contract implements transfer functions like transfer() and transferFrom() to transfer tokens from one address to another. It also implements an Approve function to allow a user to spend the token on behalf of another user. The contract also implements a mint() function to add new tokens to the total supply and a burnt() function to burn tokens. The contract emits various events like transfer_() and _mint() to keep track of the transactions.